remove Metamaps from filenames
