+++ frontend/src/Metamaps/AutoLayout.js
