+++ frontend/src/Metamaps/Debug.js
