+++ frontend/src/Metamaps/SynapseCard.js
