+++ frontend/src/Metamaps/Util.js
